受注のURL変更に対処
