+++ backend/apps/trade_flow/views/__init__.py
@@ -1,3 +1,2 @@
 from .order_view import OrderView
-from .order_update_view import OrderUpdateView
-from .order_delete_view import OrderDeleteView
+from .order_detail_view import OrderDetailView
