+++ backend/apps/trade_flow/views/order_detail_view.py
@@ -9,9 +9,10 @@ from apps.trade_flow.models.orders import Order
 from apps.trade_flow.serializers import OrderCreateSerializer
 from apps.staff_hub.permission import HasUserPermissionObject
 from apps.trade_flow.common import check_trade_flow_edit_permission
+from django.utils import timezone
 
 
-class OrderUpdateView(APIView):
+class OrderDetailView(APIView):
     permission_classes = [IsAuthenticated, HasUserPermissionObject]
 
     @extend_schema(
@@ -29,3 +30,18 @@ class OrderUpdateView(APIView):
         serializer.save()
 
         return Response(serializer.data, status=status.HTTP_200_OK)
+
+    @extend_schema(
+        tags=["order"],
+        description="注文情報を論理削除する",
+        responses={204: None}
+    )
+    def delete(self, request, pk):
+        check_trade_flow_edit_permission(request)
+
+        order = get_object_or_404(Order, pk=pk, deleted_at__isnull=True)
+
+        order.deleted_at = timezone.now()
+        order.save()
+
+        return Response(status=status.HTTP_204_NO_CONTENT)
